requirements.txt
Flask==1.0.2
flask_sqlalchemy==2.4.3
flask-restplus==0.13.0
psycopg2-binary==2.8.5
Werkzeug==0.16.1
flask-marshmallow== 0.12.0
marshmallow-sqlalchemy==0.23.1